quality-rules/requirements.txt
gspread
gspread-dataframe
pandas
openpyxl
numpy
pytz
fastapi[standard]==0.115.8
google-auth
google-api-core
google-cloud-dataplex
google-cloud-bigquery
google-cloud-datacatalog
google-cloud-storage
google-cloud-pubsub